Make sure we add owners when creating documents
Relationship creation requires Update permission
on source and destination so we need to add the
current user to a list of owners of the document
object for them to get Update permissions. We did
not have this permission issue when we used the
object_documents table because the object_documents
table did not have the same permission check as
relationships.