Check conditions for null context on the frontend
This commit changes how we check permissions on the frontend.
Up until the archiving feature we never had any conditions on
the null context, but because a global editor can edit an audit
only when it's not archived we need to add a check for permissions
on the null context as well (otherwise the editor will not be able
to edit the audit - this was the bug the commit fixed).